+++ cli/complete.go
@@ -203,6 +203,14 @@ func completeOptValue(cmd *cobra.Command, args []string, toComplete string) ([]s
 
 	var a []string
 	switch opt.(type) {
+	case options.String:
+		if opt.Key() == OptLogFile.Key() {
+			// We return the default directive, so that the shell will offer
+			// regular ol' file completion.
+			return a, cobra.ShellCompDirectiveDefault
+		}
+	case LogLevelOpt:
+		a = []string{"debug", "DEBUG", "info", "INFO", "warn", "WARN", "error", "ERROR"}
 	case FormatOpt:
 		a = stringz.Strings(format.All())
 	case options.Bool: